requirements.txt
PySide6>=6.6.0
google-genai>=1.0.0
gTTS>=2.5.0
pygame>=2.5.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
